requirements.txt
airtable-python-wrapper==0.15.1
certifi==2020.11.8
chardet==3.0.4
idna==2.10
newsapi-python==0.2.6
requests==2.25.0
urllib3==1.26.2
